ci: migrate releases to release-please
