+++ .release-please-manifest.json
@@ -1,0 +1,3 @@
+{
+  ".": "1.29.0"
+}
