--- .releaserc.json
@@ -1,29 +1,0 @@
-{
-  "branches": ["main"],
-  "plugins": [
-    "@semantic-release/commit-analyzer",
-    "@semantic-release/release-notes-generator",
-    "@semantic-release/changelog",
-    "@semantic-release/npm",
-    "@semantic-release/github",
-    "./scripts/sync-plugin-versions.mjs",
-    [
-      "@semantic-release/git",
-      {
-        "assets": [
-          "package.json",
-          "CHANGELOG.md",
-          "plugins/tartinerlabs/.codex-plugin/plugin.json",
-          "plugins/tartinerlabs/.claude-plugin/plugin.json",
-          "plugins/tartinerlabs/.cursor-plugin/plugin.json",
-          "plugins/tartinerlabs/.antigravity-plugin/plugin.json",
-          "plugins/xcode-skills/.codex-plugin/plugin.json",
-          "plugins/xcode-skills/.claude-plugin/plugin.json",
-          "plugins/xcode-skills/.cursor-plugin/plugin.json",
-          "plugins/xcode-skills/.antigravity-plugin/plugin.json"
-        ],
-        "message": "chore(release): ${nextRelease.version}"
-      }
-    ]
-  ]
-}
