+++ README.md
@@ -182,7 +182,7 @@ Each plugin lives in its own `plugins/<name>/` wrapper holding the three per-cha
 - Cursor metadata lives in `plugins/tartinerlabs/.cursor-plugin/plugin.json` and `.cursor-plugin/marketplace.json`
 - Antigravity metadata lives in `plugins/tartinerlabs/.antigravity-plugin/plugin.json`
 - The separate Xcode collection is wrapped by `plugins/xcode-skills/`, which links to the untouched `xcode-skills/` export
-- `package.json.version` is the shared source of truth across plugin manifests; semantic-release syncs manifest versions during release
+- `.release-please-manifest.json` is the shared source of truth across plugin manifests; release-please syncs manifest versions in the release PR
 
 ## Architecture
