+++ package.json
@@ -39,12 +39,9 @@
     "@biomejs/biome": "2.4.13",
     "@commitlint/cli": "20.4.1",
     "@commitlint/config-conventional": "20.4.1",
-    "@semantic-release/changelog": "6.0.3",
-    "@semantic-release/git": "10.0.1",
     "@types/node": "22.19.17",
     "husky": "9.1.7",
     "lint-staged": "16.4.0",
-    "semantic-release": "25.0.3",
     "typescript": "6.0.3"
   },
   "packageManager": "pnpm@11.14.0",
